refactor(simulation): 아크 그리드 시뮬레이터 패널 분리
