+++ src/components/simulation/ArkGridGemEditorModal.tsx
@@ -1,0 +1,193 @@
+import type { Dispatch, ReactElement, SetStateAction } from 'react';
+import type { ArkGridSlot } from '../../types/lostark';
+import {
+  ARK_GRID_GEM_OPTIONS,
+  ARK_GRID_GEM_SELECT_NUMBERS,
+  getArkGridGemState,
+  type ArkGridCoreMod,
+  type ArkGridGemEditTarget,
+  type ArkGridGemMod,
+} from './arkGridSimulatorState';
+
+interface ArkGridGemEditorModalProps {
+  arkGridSlots: readonly ArkGridSlot[];
+  arkGridMods: Record<number, ArkGridCoreMod>;
+  editingArkGridGem: ArkGridGemEditTarget | null;
+  arkGridGemDraft: Required<ArkGridGemMod> | null;
+  setEditingArkGridGem: Dispatch<SetStateAction<ArkGridGemEditTarget | null>>;
+  setArkGridGemDraft: Dispatch<SetStateAction<Required<ArkGridGemMod> | null>>;
+  updateArkGridGem: (
+    target: ArkGridGemEditTarget,
+    baseState: Required<ArkGridGemMod>,
+    patch: ArkGridGemMod,
+  ) => void;
+}
+
+export const ArkGridGemEditorModal = ({
+  arkGridSlots,
+  arkGridMods,
+  editingArkGridGem,
+  arkGridGemDraft,
+  setEditingArkGridGem,
+  setArkGridGemDraft,
+  updateArkGridGem,
+}: ArkGridGemEditorModalProps): ReactElement | null => {
+  const editingArkGridSlot = editingArkGridGem
+    ? arkGridSlots.find((slot) => slot.Index === editingArkGridGem.slotIndex)
+    : undefined;
+  const editingArkGridGemData = editingArkGridGem
+    ? editingArkGridSlot?.Gems?.find((gem) => gem.Index === editingArkGridGem.gemIndex)
+    : undefined;
+  const editingArkGridGemApiState = getArkGridGemState(editingArkGridGemData);
+  const editingArkGridGemMod = editingArkGridGem
+    ? arkGridMods[editingArkGridGem.slotIndex]?.gems?.[editingArkGridGem.gemIndex]
+    : undefined;
+  const editingArkGridGemState: Required<ArkGridGemMod> = {
+    willpower: editingArkGridGemMod?.willpower ?? editingArkGridGemApiState.willpower,
+    corePoint: editingArkGridGemMod?.corePoint ?? editingArkGridGemApiState.corePoint,
+    effects: editingArkGridGemMod?.effects ?? editingArkGridGemApiState.effects,
+  };
+  const displayedArkGridGemState = arkGridGemDraft ?? editingArkGridGemState;
+
+  if (!editingArkGridGem || !editingArkGridSlot) return null;
+
+  const closeEditor = (): void => {
+    setEditingArkGridGem(null);
+    setArkGridGemDraft(null);
+  };
+
+  return (
+    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/45 p-4 backdrop-blur-sm">
+      <div className="w-full max-w-md rounded-2xl border border-gray-200 bg-white p-4 shadow-2xl dark:border-white/10 dark:bg-gray-900">
+        <div className="mb-3 flex items-start justify-between gap-3">
+          <div className="flex items-center gap-3">
+            <div className="h-14 w-14 overflow-hidden rounded-xl border border-la-gold/40 bg-gray-100 shadow-inner dark:bg-black/30">
+              {editingArkGridGemData?.Icon ? (
+                <img src={editingArkGridGemData.Icon} alt="" className="h-full w-full object-cover" />
+              ) : (
+                <div className="h-full w-full bg-gray-100 dark:bg-white/5" />
+              )}
+            </div>
+            <div>
+              <p className="text-sm font-semibold text-gray-800 dark:text-gray-100">젬 옵션 선택</p>
+              <p className="mt-0.5 text-[11px] text-gray-400 dark:text-gray-500">
+                {editingArkGridGemData?.Grade ?? '아크 그리드'} 젬
+              </p>
+            </div>
+          </div>
+          <button
+            type="button"
+            onClick={closeEditor}
+            className="rounded-lg px-2 py-1 text-xs text-gray-400 hover:bg-gray-100 hover:text-gray-700 dark:hover:bg-white/10 dark:hover:text-gray-200"
+          >
+            닫기
+          </button>
+        </div>
+
+        <div className="grid grid-cols-2 gap-2">
+          <label className="block text-[11px] font-medium text-gray-500 dark:text-gray-400">
+            필요 의지력
+            <select
+              value={displayedArkGridGemState.willpower}
+              onChange={(event) => setArkGridGemDraft((prev) => ({
+                ...(prev ?? editingArkGridGemState),
+                willpower: Number(event.target.value),
+              }))}
+              className="mt-1 h-9 w-full rounded-lg border border-gray-200 bg-gray-50 px-2 text-xs font-semibold text-gray-700 outline-none focus:border-la-gold dark:border-white/10 dark:bg-black/20 dark:text-gray-200"
+            >
+              {ARK_GRID_GEM_SELECT_NUMBERS.map((value) => (
+                <option key={`willpower-${value}`} value={value}>{value}</option>
+              ))}
+            </select>
+          </label>
+          <label className="block text-[11px] font-medium text-gray-500 dark:text-gray-400">
+            질서/혼돈 포인트
+            <select
+              value={displayedArkGridGemState.corePoint}
+              onChange={(event) => setArkGridGemDraft((prev) => ({
+                ...(prev ?? editingArkGridGemState),
+                corePoint: Number(event.target.value),
+              }))}
+              className="mt-1 h-9 w-full rounded-lg border border-gray-200 bg-gray-50 px-2 text-xs font-semibold text-gray-700 outline-none focus:border-la-gold dark:border-white/10 dark:bg-black/20 dark:text-gray-200"
+            >
+              {ARK_GRID_GEM_SELECT_NUMBERS.map((value) => (
+                <option key={`core-point-${value}`} value={value}>{value}</option>
+              ))}
+            </select>
+          </label>
+        </div>
+
+        <div className="mt-3 grid grid-cols-1 gap-2 sm:grid-cols-2">
+          {displayedArkGridGemState.effects.map((effect, effectIndex) => (
+            <div
+              key={`${editingArkGridGem.slotIndex}-${editingArkGridGem.gemIndex}-${effectIndex}`}
+              className="rounded-xl border border-gray-200/70 bg-gray-50 p-2 dark:border-white/10 dark:bg-black/20"
+            >
+              <p className="mb-1 text-[11px] font-medium text-gray-500 dark:text-gray-400">
+                옵션 {effectIndex + 1}
+              </p>
+              <div className="grid grid-cols-[minmax(0,1fr)_74px] gap-2">
+                <select
+                  value={effect.option}
+                  onChange={(event) => setArkGridGemDraft((prev) => {
+                    const draft = prev ?? editingArkGridGemState;
+                    return {
+                      ...draft,
+                      effects: draft.effects.map((row, index) =>
+                        index === effectIndex ? { ...row, option: event.target.value } : row,
+                      ),
+                    };
+                  })}
+                  className="h-9 rounded-lg border border-gray-200 bg-white px-2 text-xs text-gray-700 outline-none focus:border-la-gold dark:border-white/10 dark:bg-gray-900 dark:text-gray-200"
+                  aria-label={`아크 그리드 젬 옵션 ${effectIndex + 1}`}
+                >
+                  {ARK_GRID_GEM_OPTIONS.map((option) => (
+                    <option key={option} value={option}>{option}</option>
+                  ))}
+                </select>
+                <select
+                  value={effect.level}
+                  onChange={(event) => setArkGridGemDraft((prev) => {
+                    const draft = prev ?? editingArkGridGemState;
+                    return {
+                      ...draft,
+                      effects: draft.effects.map((row, index) =>
+                        index === effectIndex ? { ...row, level: Number(event.target.value) } : row,
+                      ),
+                    };
+                  })}
+                  className="h-9 rounded-lg border border-gray-200 bg-white px-2 text-xs font-semibold text-gray-700 outline-none focus:border-la-gold dark:border-white/10 dark:bg-gray-900 dark:text-gray-200"
+                  aria-label={`아크 그리드 젬 옵션 ${effectIndex + 1} 레벨`}
+                >
+                  {ARK_GRID_GEM_SELECT_NUMBERS.map((value) => (
+                    <option key={`effect-${effectIndex}-level-${value}`} value={value}>Lv.{value}</option>
+                  ))}
+                </select>
+              </div>
+            </div>
+          ))}
+        </div>
+
+        <div className="mt-4 flex justify-end gap-2">
+          <button
+            type="button"
+            onClick={closeEditor}
+            className="rounded-lg border border-gray-200 px-3 py-1.5 text-xs font-semibold text-gray-500 hover:bg-gray-50 dark:border-white/10 dark:text-gray-400 dark:hover:bg-white/5"
+          >
+            취소
+          </button>
+          <button
+            type="button"
+            onClick={() => {
+              updateArkGridGem(editingArkGridGem, editingArkGridGemState, displayedArkGridGemState);
+              closeEditor();
+            }}
+            className="rounded-lg bg-la-gold px-3 py-1.5 text-xs font-bold text-gray-900 hover:bg-la-gold/90"
+          >
+            확인
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+};
